refactor(query): move join stat estimation into optimizer stats
