+++ src/query/sql/src/planner/optimizer/ir/mod.rs
@@ -43,6 +43,9 @@ pub use property::StatInfo;
 pub use property::Statistics;
 pub use stats::ColumnStat;
 pub use stats::ColumnStatSet;
+pub(crate) use stats::JoinConditionColumns;
+pub(crate) use stats::JoinKeyStatUpdate;
+pub(crate) use stats::JoinStatsEstimator;
 pub use stats::MAX_SELECTIVITY;
 pub use stats::SelectivityEstimator;
 pub use stats::UniformSampleSet;
