+++ src/query/sql/src/planner/optimizer/ir/stats/mod.rs
@@ -14,9 +14,13 @@
 
 mod column_stat;
 mod constraint;
+mod join;
 mod selectivity;
 
 pub use column_stat::*;
 pub use databend_common_statistics::UniformSampleSet;
+pub(crate) use join::JoinConditionColumns;
+pub(crate) use join::JoinKeyStatUpdate;
+pub(crate) use join::JoinStatsEstimator;
 pub use selectivity::MAX_SELECTIVITY;
 pub use selectivity::SelectivityEstimator;
